pyproject.toml
[tool.poetry]
name = "moex-python-sdk"
version = "0.1.0"
description = ""
authors = ["Stepan Andreev <andreev.s@r-mades.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pydantic = "^2.1.1"
Sphinx = "^7.1.2"
pytest-xdist = "^3.3.1"
pytz = "^2023.3.post1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
